[meta,refactor,style]: remove mentions of DBL, use non-scoped namespaces, prettier
