--- DiscordBotsList.Api/DiscordBotsList.Api.sln
@@ -1,24 +1,0 @@
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.5.2.0
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DiscordBotsList.Api", "DiscordBotsList.Api.csproj", "{90A02C5B-FD58-D268-3786-6D57221A3288}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{90A02C5B-FD58-D268-3786-6D57221A3288}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{90A02C5B-FD58-D268-3786-6D57221A3288}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{90A02C5B-FD58-D268-3786-6D57221A3288}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{90A02C5B-FD58-D268-3786-6D57221A3288}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {49537C12-C6B1-4F58-B443-C2A4361127A7}
-	EndGlobalSection
-EndGlobal
