+++ Topgg.Sdk.Api/Tests/Mocks/GetSelf.json
