+++ Topgg.Sdk.Api/Tests/Mocks/GetVote.json
