+++ Topgg.Sdk.Api/Tests/Mocks/GetVotes.json
