+++ Topgg.Sdk.Api/Widget.cs
@@ -1,7 +1,7 @@
 using System.Text.RegularExpressions;
-using DiscordBotsList.Api.Data;
+using Topgg.Sdk.Api.Data;
 
-namespace DiscordBotsList.Api
+namespace Topgg.Sdk.Api
 {
     public static partial class Widget
     {
@@ -14,7 +14,7 @@ namespace DiscordBotsList.Api
         /// <param name="type">The project's type.</param>
         /// <param name="id">The project ID.</param>
         /// <returns>The widget URL.</returns>
-        public static string Large(ProjectType type, ulong id) => $"{DiscordBotListApi.BaseURL}/widgets/large/{TypeConversionRegex().Replace(type.ToString(), "/$1").ToLower()}/{id}";
+        public static string Large(ProjectType type, ulong id) => $"{TopggApi.BaseURL}/widgets/large/{TypeConversionRegex().Replace(type.ToString(), "/$1").ToLower()}/{id}";
 
         /// <summary>
         ///     Generates a small widget URL for displaying votes.
@@ -22,7 +22,7 @@ namespace DiscordBotsList.Api
         /// <param name="type">The project's type.</param>
         /// <param name="id">The project ID.</param>
         /// <returns>The widget URL.</returns>
-        public static string Votes(ProjectType type, ulong id) => $"{DiscordBotListApi.BaseURL}/widgets/small/votes/{TypeConversionRegex().Replace(type.ToString(), "/$1").ToLower()}/{id}";
+        public static string Votes(ProjectType type, ulong id) => $"{TopggApi.BaseURL}/widgets/small/votes/{TypeConversionRegex().Replace(type.ToString(), "/$1").ToLower()}/{id}";
 
         /// <summary>
         ///     Generates a small widget URL for displaying a project's owner.
@@ -30,7 +30,7 @@ namespace DiscordBotsList.Api
         /// <param name="type">The project's type.</param>
         /// <param name="id">The project ID.</param>
         /// <returns>The widget URL.</returns>
-        public static string Owner(ProjectType type, ulong id) => $"{DiscordBotListApi.BaseURL}/widgets/small/owner/{TypeConversionRegex().Replace(type.ToString(), "/$1").ToLower()}/{id}";
+        public static string Owner(ProjectType type, ulong id) => $"{TopggApi.BaseURL}/widgets/small/owner/{TypeConversionRegex().Replace(type.ToString(), "/$1").ToLower()}/{id}";
 
         /// <summary>
         ///     Generates a small widget URL for displaying social stats.
@@ -38,6 +38,6 @@ namespace DiscordBotsList.Api
         /// <param name="type">The project's type.</param>
         /// <param name="id">The project ID.</param>
         /// <returns>The widget URL.</returns>
-        public static string Social(ProjectType type, ulong id) => $"{DiscordBotListApi.BaseURL}/widgets/small/social/{TypeConversionRegex().Replace(type.ToString(), "/$1").ToLower()}/{id}";
+        public static string Social(ProjectType type, ulong id) => $"{TopggApi.BaseURL}/widgets/small/social/{TypeConversionRegex().Replace(type.ToString(), "/$1").ToLower()}/{id}";
     }
 }
