+++ Topgg.Sdk.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
 VisualStudioVersion = 15.0.27130.2020
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DiscordBotsList.Api", "DiscordBotsList.Api\DiscordBotsList.Api.csproj", "{B20E0634-86B7-4392-A40D-93B368A68158}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Topgg.Sdk.Api", "Topgg.Sdk.Api\Topgg.Sdk.Api.csproj", "{B20E0634-86B7-4392-A40D-93B368A68158}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
